+++ resources/src/views/app/pages/commissions/index_programs.vue
@@ -92,6 +92,7 @@
 </template>
 
 <script>
+import { confirmDialog, notifications } from "@/platform";
 import { mapGetters } from 'vuex';
 import NProgress from 'nprogress';
 import PxPageHeader from "@/components/px-next/PxPageHeader.vue";
@@ -235,7 +236,7 @@ export default {
       });
     },
     confirmDelete(row) {
-      this.$bvModal.msgBoxConfirm(this.$t('Confirm_delete')).then((ok) => {
+      confirmDialog(this.$t('Confirm_delete'), { presentation: 'modal' }).then((ok) => {
         if (ok) {
           axios.delete(`commission_programs/${row.id}`).then(() => {
             this.makeToast('success', this.$t('Deleted'));
@@ -245,7 +246,7 @@ export default {
       });
     },
     makeToast(variant, msg, title = '') {
-      this.$bvToast.toast(msg, { title: title || this.$t('Notice'), variant, solid: true });
+      notifications.notify(msg, { title: title || this.$t('Notice'), variant, solid: true });
     },
   },
 };
